requirements.txt
openai-whisper
deep-translator
torch
numpy
pyinstaller
